Ending Day Now gets you the resulting recourses from production
added a way for factories to be put into production( this is scuffed tho, cuz factories can be put into production multiple time so have to fix that)

The amount of factories in production is counted and then is multiplied by the rate of recourse production per day (2 for now).

That number of "profit recourse" is added to the players current number of recourses.

as they start the next turn/day, factories are reset so they are no longer in production, and day specific variables (such as profit recourse)  is reset.