Pale System
BUG REPORT:
For some reason if a pale tile ends up on 2,1,0 (Bottom left side) , the system will stop working. I think this is due to how the grid checking system works.

However I have been able to circumvent this by making another pale Tile "pale Juice" and making it so that on "hit" it will change the tile from pale tile to pale juice. This for some reason makes it work, even the making a list and adding the tile position should do the same.

this is really weird and want to ask teach about it when I got chance but for now/playtest it should work