Fixed the loss of WF on gathering Bug
There was a bug where depending on how many recourses there were, and how many WF where allocated, some work force could be lost.

This was due to the fact that the variable that was used to re allocate WF after procedure was the same one as the one used to calculate profit.

be being super cool and awesome then decided to Tamper with that same variable so that if playeres allocated more workers than there was recourses to gather they will only get the amount left (E.g 2 recourse left on tile, 5 workers allocated will not net you 10 recourse but only 2 ).

fixed it by adding a temp var in the method